Track7Dev | Backend only accepts one email address across the entire system

+++ client/src/actions/user.js
@@ -34,41 +34,61 @@ export const getUserProfile = () => (dispatch) => {
     });
 };
 
-export const loginUser = (credentials, type) => (dispatch) => {
+export const loginUser = (credentials) => (dispatch) => {
   dispatch({ type: actionTypes.LOGIN_USER.IN_PROGRESS });
-
+  let types = ['jobseeker', 'employer'];
   credentials = sign(credentials);
   axios
-    .post(`/${type}s/login`, {token: credentials})
+    .post(`/${types[0]}s/login`, {token: credentials})
     .then((response) => {
       const { token, profile } = response.data;
-      localStorage.setItem('user', JSON.stringify({ type, token }));
+      localStorage.setItem('user', JSON.stringify({ type: types[0], token }));
       dispatch({ type: actionTypes.LOGIN_USER.SUCCESS, profile });
     })
     .catch((err) => {
-      dispatch({
-        type: actionTypes.LOGIN_USER.ERROR,
-        modalMessage: err.response.data.message,
+      axios
+      .post(`/${types[1]}s/login`, {token: credentials})
+      .then((response) => {
+        const { token, profile } = response.data;
+        localStorage.setItem('user', JSON.stringify({ type: types[1], token }));
+        dispatch({ type: actionTypes.LOGIN_USER.SUCCESS, profile });
+      })
+      .catch((err) => {
+        dispatch({
+          type: actionTypes.LOGIN_USER.ERROR,
+          modalMessage: "INVALID CREDENTIALS",
+        });
       });
     });
 };
 
-export const registerUser = (user, type) => (dispatch) => {
+export const registerUser = (user, type) => async (dispatch) => {
+  const ERROR_REGISTER = (err) => dispatch({
+    type: actionTypes.REGISTER_USER.ERROR,
+    modalMessage: err,
+  });
+
   dispatch({ type: actionTypes.REGISTER_USER.IN_PROGRESS });
-  user = jwt.sign(user, process.env.REACT_APP_ACCESS_KEY);
+
+  user = sign(user);
+ 
+  // Checking for employers with the same email
+  let userExist = (await axios.post(`/exist`, {token:user})).data.exist;
+ 
+  // Return with error if email exist in employers
+  if(await userExist) return ERROR_REGISTER("EMAIL IS ALREADY IN USE");
+
+  // If email has never been used proceed with registration
   axios
-    .post(`${type}s/register`, {token: user})
-    .then((response) => {
-      const { token, profile } = response.data;
-      localStorage.setItem('user', JSON.stringify({ type, token }));
-      dispatch({ type: actionTypes.REGISTER_USER.SUCCESS, profile });
-    })
-    .catch((err) => {
-      dispatch({
-        type: actionTypes.REGISTER_USER.ERROR,
-        modalMessage: err.response.data.message,
-      });
-    });
+  .post(`${type}s/register`, {token: user})
+  .then((response) => {
+    const { token, profile } = response.data;
+    localStorage.setItem('user', JSON.stringify({ type, token }));
+    dispatch({ type: actionTypes.REGISTER_USER.SUCCESS, profile });
+  })
+  .catch((err) => ERROR_REGISTER(err.response.data.message));
+    
+    
 };
 
 export const updateUserProfile = updatedInfo => (dispatch) => {
@@ -88,7 +108,7 @@ export const updateUserProfile = updatedInfo => (dispatch) => {
     .catch((err) => {
       dispatch({
         type: actionTypes.UPDATE_USER_PROFILE.ERROR,
-        modalMessage: err.response.data.message,
+        modalMessage: err.message,
       });
     });
 };
@@ -97,7 +117,7 @@ export const updateUserPassword = updatedInfo => (dispatch) => {
   dispatch({ type: actionTypes.UPDATE_USER_PROFILE.IN_PROGRESS });
 
   const user = JSON.parse(localStorage.getItem('user'));
-  updatedInfo = jwt.sign(updatedInfo, process.env.REACT_APP_ACCESS_KEY);
+  updatedInfo = sign(updatedInfo);
   const requestOptions = { // send with get on protected routes
     headers: {
       Authorization: `Bearer ${user.token}`,
